fix: use const for config
